vp8l_dec: don't truncate the scratch-row size
AllocateInternalBuffers32b(): 'cache_top_pixels' is (uint16_t)final_width,
while 'cache_pixels' just below uses (uint64_t)final_width. A no-op only
because WEBP_MAX_DIMENSION is 16383, on a row VP8LInverseTransform() writes.
=> drop the cast.

+++ src/dec/vp8l_dec.c
@@ -1644,7 +1644,7 @@ static int AllocateInternalBuffers32b(VP8LDecoder* const dec, int final_width) {
   const uint64_t num_pixels = (uint64_t)dec->width * dec->height;
   // Scratch buffer corresponding to top-prediction row for transforming the
   // first row in the row-blocks. Not needed for paletted alpha.
-  const uint64_t cache_top_pixels = (uint16_t)final_width;
+  const uint64_t cache_top_pixels = final_width;
   // Scratch buffer for temporary BGRA storage. Not needed for paletted alpha.
   const uint64_t cache_pixels = (uint64_t)final_width * NUM_ARGB_CACHE_ROWS;
   // Scratch buffer to accumulate RGBA values (hence 4*)for YUV conversion.
